requirements.txt
elasticsearch==6.3.1
PyYAML==3.13
speedtest-cli==2.0.2
urllib3==1.24.1
